IT'S DONE!!! IT'S FINALLY DONE!!!
Not finally:  in fact, it just works and I need to check code smells & convention in sonar. Probably it is needed to create spetial method "again play" to clean Main a little bit. There is also an option to combine Points and PotentialPoints methods together, but Idk how to do it right now.